api/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.6.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
xgboost==2.0.3
pyjwt==2.8.0
langchain
langchain-openai
langchain-groq
openai==1.30.5
python-dotenv==1.0.1
gunicorn==22.0.0

